+++ pkg/goweb/server_test.go
@@ -1,4 +1,4 @@
-package webserver
+package goweb
 
 import (
 	"errors"
